Admin form exceptions handles and form validation

+++ Bankify/Form_Admin_LoginAccount.cs
@@ -15,70 +15,90 @@ namespace Bankify
         public Form_Admin_LoginAccount()
         {
             InitializeComponent();
+            CenterToScreen();
         }
 
         private void Form_Admin_LoginAccount_Load(object sender, EventArgs e)
         {
             // TODO: This line of code loads data into the 'bank_dbDataSet2.LoginAccount' table. You can move, or remove it, as needed.
             this.loginAccountTableAdapter.Fill(this.bank_dbDataSet2.LoginAccount);
-
+            accountype_CB.DropDownStyle = ComboBoxStyle.DropDownList;
         }
 
         private void AddLA_Click(object sender, EventArgs e)
         {
-            LoginAccount loginAccount = new LoginAccount();
-            using (var db = new Bank_dbEntities())
+            if (String.IsNullOrWhiteSpace(username_TB.Text) || String.IsNullOrWhiteSpace(password_Tb.Text))
             {
-                try
+                MessageBox.Show("Username sau Parola invalida");
+            }
+            else
+            {
+                LoginAccount loginAccount = new LoginAccount();
+                using (var db = new Bank_dbEntities())
                 {
                     try
                     {
-                        loginAccount.login_username = username_TB.Text;
-                        loginAccount.login_password = password_Tb.Text;
-                        loginAccount.account_type = accountype_CB.Text;
                         try
                         {
-                            db.LoginAccount.Add(loginAccount);
-                            db.SaveChanges();
-                            MessageBox.Show("Adaugare reusita!");
+                            loginAccount.login_username = username_TB.Text;
+                            loginAccount.login_password = password_Tb.Text;
+                            loginAccount.account_type = accountype_CB.Text;
+                            try
+                            {
+                                db.LoginAccount.Add(loginAccount);
+                                db.SaveChanges();
+                                MessageBox.Show("Adaugare reusita!");
 
-                            this.loginAccountTableAdapter.Fill(this.bank_dbDataSet2.LoginAccount);
+                                this.loginAccountTableAdapter.Fill(this.bank_dbDataSet2.LoginAccount);
+
+                            }
+                            catch
+                            {
+                                MessageBox.Show("Adaugarea a esuat-------!");
+                            }
 
                         }
                         catch
                         {
-                            MessageBox.Show("Adaugarea a esuat-------!");
+                            MessageBox.Show("Adaugarea a esuat!");
                         }
 
                     }
                     catch
                     {
-                        MessageBox.Show("Adaugarea a esuat!");
+
                     }
-
-                }
-                catch
-                {
-
                 }
             }
         }
 
         private void editLA_Click(object sender, EventArgs e)
         {
-            int login_account_id = int.Parse(dataGridView_LA.CurrentRow.Cells[0].Value.ToString());
-            using (var db = new Bank_dbEntities())
+            if (String.IsNullOrWhiteSpace(username_TB.Text) || String.IsNullOrWhiteSpace(password_Tb.Text))
             {
-                var login_account = (from c in db.LoginAccount
-                                     where c.login_id == login_account_id
-                                     select c).First();
-                this.loginAccountTableAdapter.Update(username_TB.Text, password_Tb.Text, accountype_CB.Text, login_account.login_id,
-                    login_account.login_username, login_account.login_password, login_account.account_type);
-                db.SaveChanges();
-                MessageBox.Show("Modificare reusita!");
-                this.loginAccountTableAdapter.Fill(this.bank_dbDataSet2.LoginAccount);
-
-
+                MessageBox.Show("Username sau Parola invalida");
+            }
+            else
+            {
+                try
+                {
+                    int login_account_id = int.Parse(dataGridView_LA.CurrentRow.Cells[0].Value.ToString());
+                    using (var db = new Bank_dbEntities())
+                    {
+                        var login_account = (from c in db.LoginAccount
+                                             where c.login_id == login_account_id
+                                             select c).First();
+                        this.loginAccountTableAdapter.Update(username_TB.Text, password_Tb.Text, accountype_CB.Text, login_account.login_id,
+                            login_account.login_username, login_account.login_password, login_account.account_type);
+                        db.SaveChanges();
+                        MessageBox.Show("Modificare reusita!");
+                        this.loginAccountTableAdapter.Fill(this.bank_dbDataSet2.LoginAccount);
+                    }
+                }
+                catch (InvalidOperationException ex1)
+                {
+                    MessageBox.Show("Date introduse gresit!");
+                }
             }
         }
 
@@ -99,24 +119,29 @@ namespace Bankify
         private void deleteLA_Click(object sender, EventArgs e)
         {
             int loginaccount_id = int.Parse(dataGridView_LA.CurrentRow.Cells[0].Value.ToString());
-            using (var db = new Bank_dbEntities())
+            ConfirmForm confirmForm = new ConfirmForm("Doriti sa stergeti acest login account?");
+            confirmForm.ShowDialog();
+            if (ConfirmForm.option == 1)
             {
-
-                try
+                using (var db = new Bank_dbEntities())
                 {
-                    var loginAccount = (from c in db.LoginAccount
-                                  where c.login_id == loginaccount_id
-                                  select c).First();
-                    db.LoginAccount.Remove(loginAccount);
-                    db.SaveChanges();
-                    dataGridView_LA.Rows.RemoveAt(this.dataGridView_LA.CurrentCell.RowIndex);
-                    MessageBox.Show("Stergere reusita!");
-                    this.loginAccountTableAdapter.Fill(this.bank_dbDataSet2.LoginAccount);
 
-                }
-                catch
-                {
-                    MessageBox.Show("Stergerea a esuat!");
+                    try
+                    {
+                        var loginAccount = (from c in db.LoginAccount
+                                            where c.login_id == loginaccount_id
+                                            select c).First();
+                        db.LoginAccount.Remove(loginAccount);
+                        db.SaveChanges();
+                        dataGridView_LA.Rows.RemoveAt(this.dataGridView_LA.CurrentCell.RowIndex);
+                        MessageBox.Show("Stergere reusita!");
+                        this.loginAccountTableAdapter.Fill(this.bank_dbDataSet2.LoginAccount);
+
+                    }
+                    catch
+                    {
+                        MessageBox.Show("Stergerea a esuat!");
+                    }
                 }
             }
         }
